Added scraper, and implemented its use in serv.js
scraper.js gathers data from the database, bins it,
and stores all levels above a threshold in a file for
serv.js to use.

serv.js currently reads that file in, and does not
request anything from the database. It's currently
only using the stored data.

Future plans: allow serv.js to do both, binning to a
new smaller binnedData object with new data from the
database when that data is requested.